Move ItunesApiService to Lib.Services
