+++ src/Lib.Services/Services/ItunesApiService/ItunesApiServiceExtensions.cs
@@ -1,7 +1,7 @@
 using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace MuzakBot.App.Services;
+namespace MuzakBot.Lib.Services;
 
 /// <summary>
 /// Extension methods for configuring the <see cref="ItunesApiService"/> in the dependency injection container.
