+++ src/Lib.Services/Services/ItunesApiService/interfaces/IItunesApiService.cs
@@ -1,6 +1,6 @@
 using MuzakBot.Lib.Models.Itunes;
 
-namespace MuzakBot.App.Services;
+namespace MuzakBot.Lib.Services;
 
 public interface IItunesApiService : IDisposable
 {
